Move buff override logic from buffwindow to raidbuffwindow
